git update 19 nov 2012
fontconfig: c20ac78b01df3f0919352bba16b5b48b3b5d4d6d
libxcb: 76a2166de9c80b35f987fdc3f3a228bafa0de94e
mesa: ddb901fbf4489ffcd85d3320f23913eb1d4fbdfe
pixman: 44dd746bb68625b2f6be77c3f80292b45defe9d7
xserver: 6a6c3afe71ac82a93d9fd0034dd5bbdcf0eae1ea
xkeyboard-config: 709e05c069428236ca2567e784c9971eecc8ca50

+++ pixman/pixman/pixman-arm-simd.c
@@ -56,7 +56,7 @@ pixman_composite_add_8_8_asm_armv6 (int32_t  width,
 	/* ensure both src and dst are properly aligned before doing 32 bit reads
 	 * we'll stay in this loop if src and dst have differing alignments
 	 */
-	while (w && (((unsigned long)dst & 3) || ((unsigned long)src & 3)))
+	while (w && (((uintptr_t)dst & 3) || ((uintptr_t)src & 3)))
 	{
 	    s = *src;
 	    d = *dst;
